requirements.txt
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.1
